Analytics into its own namespace now.
