+++ analytics.cpp
@@ -7,18 +7,16 @@
 // found online at https://opensource.org/licenses/MIT.
 //
 
-#include "analytics.h"
-
-#include "json.hpp"
-
 #include <chrono>
 #include <cstring>
 #include <iostream>
 #include <string>
 #include <thread>
 
+#include "analytics.hpp"
 #include "http-curl.hpp"
 #include "http-none.hpp"
+#include "json.hpp"
 
 // To keep this simple, we include the entire implementation in this one
 // C++ file.  Organizationally we'd probably have rather split this up,
@@ -28,430 +26,416 @@
 using json = nlohmann::json;
 
 namespace segment {
+namespace analytics {
 
-Event::Event(EventType type)
-    : type(type)
-{
-}
+    Event::Event(EventType type)
+        : type(type)
+    {
+    }
 
-Event::~Event()
-{
-    properties.clear();
-}
+    Event::~Event()
+    {
+        properties.clear();
+    }
 
-std::string Event::Type() const
-{
-    switch (type) {
-    case EVENT_TYPE_IDENTIFY:
-        return "identify";
-        break;
-    case EVENT_TYPE_SCREEN:
-        return "screen";
-        break;
-    case EVENT_TYPE_PAGE:
-        return "page";
-        break;
-    case EVENT_TYPE_ALIAS:
-        return "alias";
-        break;
-    case EVENT_TYPE_TRACK:
-        return "track";
-        break;
-    default:
-        throw std::invalid_argument("Invalid event type");
-    }
-}
-
-void to_json(json& j, const Event& ev)
-{
-    j["type"] = ev.Type();
-    if (ev.event != "") {
-        j["event"] = ev.event;
-    }
-    if (ev.userId != "") {
-        j["userId"] = ev.userId;
-    }
-    if (ev.groupId != "") {
-        j["groupId"] = ev.groupId;
-    }
-    if (ev.anonymousId != "") {
-        j["anonymousId"] = ev.anonymousId;
-    }
-    if (ev.previousId != "") {
-        j["previousId"] = ev.previousId;
-    }
-    if (!ev.properties.empty()) {
-        // XXX: Reading the API docs, I think this should be true for group
-        // as well?
-        if (ev.Type() == "identify") {
-            j["traits"] = ev.properties;
-        } else {
-            j["properties"] = ev.properties;
+    std::string Event::Type() const
+    {
+        switch (type) {
+        case EVENT_TYPE_IDENTIFY:
+            return "identify";
+            break;
+        case EVENT_TYPE_SCREEN:
+            return "screen";
+            break;
+        case EVENT_TYPE_PAGE:
+            return "page";
+            break;
+        case EVENT_TYPE_ALIAS:
+            return "alias";
+            break;
+        case EVENT_TYPE_TRACK:
+            return "track";
+            break;
+        default:
+            throw std::invalid_argument("Invalid event type");
         }
     }
-}
 
-void to_json(json& j, std::shared_ptr<Event> ev)
-{
-    return (to_json(j, *ev));
-}
-
-std::string Event::Serialize() const
-{
-    json j;
-    to_json(j, *this);
-    return (j.dump());
-}
-
-Analytics::Analytics(std::string writeKey)
-    : writeKey(writeKey)
-{
-    host = "https://api.segment.io";
-    Handler = std::make_shared<segment::http::HttpHandlerCurl>();
-    thr = std::thread(worker, this);
-    MaxRetries = 5;
-    RetryInterval = std::chrono::seconds(1);
-    shutdown = false;
-    FlushCount = 250;
-    FlushSize = 500 * 1024;
-    FlushInterval = std::chrono::seconds(10);
-    needFlush = false;
-    wakeTime = std::chrono::time_point<std::chrono::system_clock>::max();
-}
-
-Analytics::Analytics(std::string writeKey, std::string host)
-    : writeKey(writeKey)
-    , host(host)
-{
-    Handler = std::make_shared<segment::http::HttpHandlerCurl>();
-    thr = std::thread(worker, this);
-    MaxRetries = 5;
-    RetryInterval = std::chrono::seconds(1);
-    shutdown = false;
-    FlushCount = 250;
-    FlushSize = 500 * 1024;
-    FlushInterval = std::chrono::seconds(10);
-    needFlush = false;
-    wakeTime = std::chrono::time_point<std::chrono::system_clock>::max();
-}
-
-Analytics::~Analytics()
-{
-    FlushWait();
-    std::unique_lock<std::mutex> lk(this->lock);
-    this->shutdown = true;
-    this->cv.notify_all();
-    lk.unlock();
-
-    thr.join();
-}
-
-void Analytics::FlushWait()
-{
-    std::unique_lock<std::mutex> lk(this->lock);
-
-    // NB: If an event has been taken off the queue and is being
-    // processed, then the lock will be held, preventing us from
-    // executing this check.
-    while (!events.empty()) {
-        needFlush = true;
-        cv.notify_all();
-        cv.wait(lk);
-    }
-}
-
-void Analytics::Flush()
-{
-    std::unique_lock<std::mutex> lk(this->lock);
-    needFlush = true;
-    cv.notify_all();
-}
-
-void Analytics::Scrub()
-{
-    std::unique_lock<std::mutex> lk(this->lock);
-    events.clear();
-    cv.notify_all();
-}
-
-void Analytics::Track(std::string userId, std::string event)
-{
-    this->Track(userId, event, {});
-}
-
-void Analytics::Track(std::string userId, std::string event, std::map<std::string, std::string> properties)
-{
-    auto e = std::make_shared<Event>(EVENT_TYPE_TRACK);
-    e->userId = userId;
-    e->event = event;
-    e->properties = properties; // TODO: probably need to copy this
-
-    this->queueEvent(e);
-}
-
-void Analytics::Identify(std::string userId)
-{
-    this->Identify(userId, {});
-}
-
-void Analytics::Identify(std::string userId, std::map<std::string, std::string> traits)
-{
-    auto e = std::make_shared<Event>(EVENT_TYPE_IDENTIFY);
-    e->userId = userId;
-    e->properties = traits;
-
-    this->queueEvent(e);
-}
-
-void Analytics::Page(std::string event, std::string userId)
-{
-    this->Page(event, userId, {});
-}
-
-void Analytics::Page(std::string event, std::string userId, std::map<std::string, std::string> properties)
-{
-    auto e = std::make_shared<Event>(EVENT_TYPE_PAGE);
-    e->userId = userId;
-    e->properties = properties;
-
-    this->queueEvent(e);
-}
-
-void Analytics::Screen(std::string event, std::string userId)
-{
-    this->Screen(event, userId, {});
-}
-
-void Analytics::Screen(std::string event, std::string userId, std::map<std::string, std::string> properties)
-{
-    auto e = std::make_shared<Event>(EVENT_TYPE_SCREEN);
-    e->userId = userId;
-    e->properties = properties;
-
-    queueEvent(e);
-}
-
-void Analytics::Alias(std::string previousId, std::string userId)
-{
-    auto e = std::make_shared<Event>(EVENT_TYPE_ALIAS);
-    e->userId = userId;
-    e->previousId = previousId;
-
-    queueEvent(e);
-}
-
-void Analytics::Group(std::string groupId)
-{
-    this->Group(groupId);
-}
-
-void Analytics::Group(std::string groupId, std::map<std::string, std::string> properties)
-{
-    auto e = std::make_shared<Event>(EVENT_TYPE_GROUP);
-    e->groupId = groupId;
-    e->properties = properties;
-
-    queueEvent(e);
-}
-
-// This implementation of base64 is taken from StackOverflow:
-// https://stackoverflow.com/questions/180947/base64-decode-snippet-in-c
-// We use it here perform the base64 encode for user authentication with
-// Basic Auth, freeing the transport providers from dealing with this.
-// (Note that libcurl has this builtin though.)
-static std::string base64_encode(const std::string& in)
-{
-    std::string out;
-
-    int val = 0, valb = -6;
-    for (unsigned char c : in) {
-        val = (val << 8) + c;
-        valb += 8;
-        while (valb >= 0) {
-            out.push_back("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/"[(val >> valb) & 0x3F]);
-            valb -= 6;
+    void to_json(json& j, const Event& ev)
+    {
+        j["type"] = ev.Type();
+        if (ev.event != "") {
+            j["event"] = ev.event;
+        }
+        if (ev.userId != "") {
+            j["userId"] = ev.userId;
+        }
+        if (ev.groupId != "") {
+            j["groupId"] = ev.groupId;
+        }
+        if (ev.anonymousId != "") {
+            j["anonymousId"] = ev.anonymousId;
+        }
+        if (ev.previousId != "") {
+            j["previousId"] = ev.previousId;
+        }
+        if (!ev.properties.empty()) {
+            // XXX: Reading the API docs, I think this should be true for group
+            // as well?
+            if (ev.Type() == "identify") {
+                j["traits"] = ev.properties;
+            } else {
+                j["properties"] = ev.properties;
+            }
         }
     }
-    if (valb > -6)
-        out.push_back("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/"[((val << 8) >> (valb + 8)) & 0x3F]);
-    while (out.size() % 4)
-        out.push_back('=');
-    return out;
-}
 
-void Analytics::sendBatch()
-{
-    segment::http::HttpRequest req;
-    json body;
-    body["batch"] = batch;
-    // XXX add default context or integrations?
-
-    req.Method = "POST";
-    req.URL = this->host + "/v1/batch";
-    req.Headers["User-Agent"] = "AnalyticsCPP/0.1";
-    // Note: libcurl could do this for us, but for other transports
-    // we do it here -- keeping the transports as unaware as possible.
-    req.Headers["Authorization"] = "Basic " + base64_encode(this->writeKey);
-    req.Headers["Content-Type"] = "application/json";
-    req.Headers["Accept"] = "application/json";
-    req.Body = body.dump();
-
-    auto resp = this->Handler->Handle(req);
-}
-
-void Analytics::sendEvent(std::shared_ptr<Event> e)
-{
-    segment::http::HttpRequest req;
-
-    req.Method = "POST";
-    req.URL = this->host + "/v1/" + e->Type();
-    req.Headers["User-Agent"] = "AnalyticsCPP/0.1";
-    // Note: libcurl could do this for us, but for other transports
-    // we do it here -- keeping the transports as unaware as possible.
-    req.Headers["Authorization"] = "Basic " + base64_encode(this->writeKey);
-    req.Headers["Content-Type"] = "application/json";
-    req.Headers["Accept"] = "application/json";
-    req.Body = e->Serialize();
-
-    auto resp = this->Handler->Handle(req);
-}
-
-void Analytics::queueEvent(std::shared_ptr<Event> ev)
-{
-    std::lock_guard<std::mutex> lk(lock);
-    events.push_back(ev);
-    if (events.size() == 1) {
-        flushTime = std::chrono::system_clock::now() + FlushInterval;
-        if (flushTime < wakeTime) {
-            wakeTime = flushTime;
-        }
+    void to_json(json& j, std::shared_ptr<Event> ev)
+    {
+        return (to_json(j, *ev));
     }
-    cv.notify_all();
-}
 
-void Analytics::processQueue()
-{
-    int fails = 0;
-    bool ok;
-    std::string reason;
-    std::deque<std::shared_ptr<Event>> notifyq;
+    std::string Event::Serialize() const
+    {
+        json j;
+        to_json(j, *this);
+        return (j.dump());
+    }
 
-    for (;;) {
+    Analytics::Analytics(std::string writeKey)
+        : writeKey(writeKey)
+    {
+        host = "https://api.segment.io";
+        Handler = std::make_shared<segment::http::HttpHandlerCurl>();
+        thr = std::thread(worker, this);
+        MaxRetries = 5;
+        RetryInterval = std::chrono::seconds(1);
+        shutdown = false;
+        FlushCount = 250;
+        FlushSize = 500 * 1024;
+        FlushInterval = std::chrono::seconds(10);
+        needFlush = false;
+        wakeTime = std::chrono::time_point<std::chrono::system_clock>::max();
+    }
+
+    Analytics::Analytics(std::string writeKey, std::string host)
+        : writeKey(writeKey)
+        , host(host)
+    {
+        Handler = std::make_shared<segment::http::HttpHandlerCurl>();
+        thr = std::thread(worker, this);
+        MaxRetries = 5;
+        RetryInterval = std::chrono::seconds(1);
+        shutdown = false;
+        FlushCount = 250;
+        FlushSize = 500 * 1024;
+        FlushInterval = std::chrono::seconds(10);
+        needFlush = false;
+        wakeTime = std::chrono::time_point<std::chrono::system_clock>::max();
+    }
+
+    Analytics::~Analytics()
+    {
+        FlushWait();
+        std::unique_lock<std::mutex> lk(this->lock);
+        this->shutdown = true;
+        this->cv.notify_all();
+        lk.unlock();
+
+        thr.join();
+    }
+
+    void Analytics::FlushWait()
+    {
         std::unique_lock<std::mutex> lk(this->lock);
 
-        if (events.empty() && batch.empty()) {
-            // Reset failure count so we start with a clean slate.
-            // Otherwise we could have a failure hours earlier that
-            // allows only one failed post hours later.
-            fails = 0;
-            wakeTime = std::chrono::time_point<std::chrono::system_clock>::max();
-
-            // We might have a flusher waiting
-            cv.notify_all();
-
-            // We only shut down if the queue was empty.  To force
-            // a shutdown without draining, just clear the queue
-            // independently.
-            if (shutdown) {
-                return;
-            }
-
-            cv.wait(lk);
-            continue;
-        }
-
-        // Gather up new items into the batch, assuming that the batch
-        // is not already full.
-        while ((!events.empty()) && (batch.size() < FlushCount)) {
-            json j;
-
-            // Try adding an event to the batch and checking the
-            // serialization...  This is pretty ineffecient as we
-            // serialize the objects multiple times, but it's easier
-            // to understand.  Later look at saving the last size,
-            // and only adding the size of the serialized event.
-
-            auto ev = events.front();
-            batch.push_back(ev);
-            j["batch"] = batch;
-            if (j.dump().size() >= FlushSize) {
-                batch.pop_back(); // remove what we just added.
-                needFlush = true;
-                break;
-            }
-            // We inclucded this event, so remove it from the queue
-            // (it is already inthe batch.)
-            events.pop_front();
-        }
-
-        // WWe hit the limit.
-        if (batch.size() >= FlushCount) {
+        // NB: If an event has been taken off the queue and is being
+        // processed, then the lock will be held, preventing us from
+        // executing this check.
+        while (!events.empty()) {
             needFlush = true;
+            cv.notify_all();
+            cv.wait(lk);
         }
+    }
 
-        auto now = std::chrono::system_clock::now();
+    void Analytics::Flush()
+    {
+        std::unique_lock<std::mutex> lk(this->lock);
+        needFlush = true;
+        cv.notify_all();
+    }
 
-        if ((!needFlush) && (now < wakeTime)) {
-            cv.wait_until(lk, wakeTime);
-            continue;
+    void Analytics::Scrub()
+    {
+        std::unique_lock<std::mutex> lk(this->lock);
+        events.clear();
+        cv.notify_all();
+    }
+
+    void Analytics::Track(std::string userId, std::string event)
+    {
+        this->Track(userId, event, {});
+    }
+
+    void Analytics::Track(std::string userId, std::string event, std::map<std::string, std::string> properties)
+    {
+        auto e = std::make_shared<Event>(EVENT_TYPE_TRACK);
+        e->userId = userId;
+        e->event = event;
+        e->properties = properties; // TODO: probably need to copy this
+
+        this->queueEvent(e);
+    }
+
+    void Analytics::Identify(std::string userId)
+    {
+        this->Identify(userId, {});
+    }
+
+    void Analytics::Identify(std::string userId, std::map<std::string, std::string> traits)
+    {
+        auto e = std::make_shared<Event>(EVENT_TYPE_IDENTIFY);
+        e->userId = userId;
+        e->properties = traits;
+
+        this->queueEvent(e);
+    }
+
+    void Analytics::Page(std::string event, std::string userId)
+    {
+        this->Page(event, userId, {});
+    }
+
+    void Analytics::Page(std::string event, std::string userId, std::map<std::string, std::string> properties)
+    {
+        auto e = std::make_shared<Event>(EVENT_TYPE_PAGE);
+        e->userId = userId;
+        e->properties = properties;
+
+        this->queueEvent(e);
+    }
+
+    void Analytics::Screen(std::string event, std::string userId)
+    {
+        this->Screen(event, userId, {});
+    }
+
+    void Analytics::Screen(std::string event, std::string userId, std::map<std::string, std::string> properties)
+    {
+        auto e = std::make_shared<Event>(EVENT_TYPE_SCREEN);
+        e->userId = userId;
+        e->properties = properties;
+
+        queueEvent(e);
+    }
+
+    void Analytics::Alias(std::string previousId, std::string userId)
+    {
+        auto e = std::make_shared<Event>(EVENT_TYPE_ALIAS);
+        e->userId = userId;
+        e->previousId = previousId;
+
+        queueEvent(e);
+    }
+
+    void Analytics::Group(std::string groupId)
+    {
+        this->Group(groupId);
+    }
+
+    void Analytics::Group(std::string groupId, std::map<std::string, std::string> properties)
+    {
+        auto e = std::make_shared<Event>(EVENT_TYPE_GROUP);
+        e->groupId = groupId;
+        e->properties = properties;
+
+        queueEvent(e);
+    }
+
+    // This implementation of base64 is taken from StackOverflow:
+    // https://stackoverflow.com/questions/180947/base64-decode-snippet-in-c
+    // We use it here perform the base64 encode for user authentication with
+    // Basic Auth, freeing the transport providers from dealing with this.
+    // (Note that libcurl has this builtin though.)
+    static std::string base64_encode(const std::string& in)
+    {
+        std::string out;
+
+        int val = 0, valb = -6;
+        for (unsigned char c : in) {
+            val = (val << 8) + c;
+            valb += 8;
+            while (valb >= 0) {
+                out.push_back("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/"[(val >> valb) & 0x3F]);
+                valb -= 6;
+            }
         }
+        if (valb > -6)
+            out.push_back("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/"[((val << 8) >> (valb + 8)) & 0x3F]);
+        while (out.size() % 4)
+            out.push_back('=');
+        return out;
+    }
 
-        // We're trying to flush, so clear our "need".
-        needFlush = false;
+    void Analytics::sendBatch()
+    {
+        segment::http::HttpRequest req;
+        json body;
+        body["batch"] = batch;
+        // XXX add default context or integrations?
 
-        try {
-            sendBatch();
-            ok = true;
-            fails = false;
-        } catch (std::exception& e) {
-            if (fails < MaxRetries) {
-                // Something bad happened.  Let's wait a bit and
-                // try again later.  We return this even to the
-                // front.
-                fails++;
-                retryTime = now + RetryInterval;
-                if (retryTime < wakeTime) {
-                    wakeTime = retryTime;
+        req.Method = "POST";
+        req.URL = this->host + "/v1/batch";
+        req.Headers["User-Agent"] = "AnalyticsCPP/0.1";
+        // Note: libcurl could do this for us, but for other transports
+        // we do it here -- keeping the transports as unaware as possible.
+        req.Headers["Authorization"] = "Basic " + base64_encode(this->writeKey);
+        req.Headers["Content-Type"] = "application/json";
+        req.Headers["Accept"] = "application/json";
+        req.Body = body.dump();
+
+        auto resp = this->Handler->Handle(req);
+    }
+
+    void Analytics::queueEvent(std::shared_ptr<Event> ev)
+    {
+        std::lock_guard<std::mutex> lk(lock);
+        events.push_back(ev);
+        if (events.size() == 1) {
+            flushTime = std::chrono::system_clock::now() + FlushInterval;
+            if (flushTime < wakeTime) {
+                wakeTime = flushTime;
+            }
+        }
+        cv.notify_all();
+    }
+
+    void Analytics::processQueue()
+    {
+        int fails = 0;
+        bool ok;
+        std::string reason;
+        std::deque<std::shared_ptr<Event>> notifyq;
+
+        for (;;) {
+            std::unique_lock<std::mutex> lk(this->lock);
+
+            if (events.empty() && batch.empty()) {
+                // Reset failure count so we start with a clean slate.
+                // Otherwise we could have a failure hours earlier that
+                // allows only one failed post hours later.
+                fails = 0;
+                wakeTime = std::chrono::time_point<std::chrono::system_clock>::max();
+
+                // We might have a flusher waiting
+                cv.notify_all();
+
+                // We only shut down if the queue was empty.  To force
+                // a shutdown without draining, just clear the queue
+                // independently.
+                if (shutdown) {
+                    return;
                 }
+
+                cv.wait(lk);
+                continue;
+            }
+
+            // Gather up new items into the batch, assuming that the batch
+            // is not already full.
+            while ((!events.empty()) && (batch.size() < FlushCount)) {
+                json j;
+
+                // Try adding an event to the batch and checking the
+                // serialization...  This is pretty ineffecient as we
+                // serialize the objects multiple times, but it's easier
+                // to understand.  Later look at saving the last size,
+                // and only adding the size of the serialized event.
+
+                auto ev = events.front();
+                batch.push_back(ev);
+                j["batch"] = batch;
+                if (j.dump().size() >= FlushSize) {
+                    batch.pop_back(); // remove what we just added.
+                    needFlush = true;
+                    break;
+                }
+                // We inclucded this event, so remove it from the queue
+                // (it is already inthe batch.)
+                events.pop_front();
+            }
+
+            // WWe hit the limit.
+            if (batch.size() >= FlushCount) {
+                needFlush = true;
+            }
+
+            auto now = std::chrono::system_clock::now();
+
+            if ((!needFlush) && (now < wakeTime)) {
                 cv.wait_until(lk, wakeTime);
                 continue;
             }
-            ok = false;
-            reason = e.what();
-            // We intentionally have chosen not to reset the failure
-            // count.  Which means if we wind up failing to send one
-            // event after maxtries, we only try each of the following
-            // one time, until either the queue is empty or we have
-            // a success.
-        }
-        auto cb = Callback;
-        notifyq = batch;
-        batch.clear();
-        lk.unlock();
 
-        while (!notifyq.empty()) {
-            auto ev = notifyq.front();
-            notifyq.pop_front();
+            // We're trying to flush, so clear our "need".
+            needFlush = false;
+
             try {
-                if (cb != nullptr) {
-                    if (ok) {
-                        cb->Success(ev);
-                    } else {
-                        cb->Failure(ev, reason);
-                    }
-                }
+                sendBatch();
+                ok = true;
+                fails = false;
             } catch (std::exception& e) {
-                // User supplied callback code failed.  There isn't really
-                // anything else we can do.  Muddle on.  This prevents a
-                // failure there from silently causing the processing thread
-                // to stop functioning.
+                if (fails < MaxRetries) {
+                    // Something bad happened.  Let's wait a bit and
+                    // try again later.  We return this even to the
+                    // front.
+                    fails++;
+                    retryTime = now + RetryInterval;
+                    if (retryTime < wakeTime) {
+                        wakeTime = retryTime;
+                    }
+                    cv.wait_until(lk, wakeTime);
+                    continue;
+                }
+                ok = false;
+                reason = e.what();
+                // We intentionally have chosen not to reset the failure
+                // count.  Which means if we wind up failing to send one
+                // event after maxtries, we only try each of the following
+                // one time, until either the queue is empty or we have
+                // a success.
+            }
+            auto cb = Callback;
+            notifyq = batch;
+            batch.clear();
+            lk.unlock();
+
+            while (!notifyq.empty()) {
+                auto ev = notifyq.front();
+                notifyq.pop_front();
+                try {
+                    if (cb != nullptr) {
+                        if (ok) {
+                            cb->Success(ev);
+                        } else {
+                            cb->Failure(ev, reason);
+                        }
+                    }
+                } catch (std::exception& e) {
+                    // User supplied callback code failed.  There isn't really
+                    // anything else we can do.  Muddle on.  This prevents a
+                    // failure there from silently causing the processing thread
+                    // to stop functioning.
+                }
             }
         }
     }
-}
 
-void Analytics::worker(Analytics* self)
-{
-    self->processQueue();
-}
-}
+    void Analytics::worker(Analytics* self)
+    {
+        self->processQueue();
+    }
+
+} // namespace analytics
+} // namespace segment
